+++ Event/xAOD/xAODMissingET/xAODMissingET/versions/MissingETAuxContainer_v1.h
@@ -36,9 +36,9 @@ namespace xAOD
      *         exactly the same spelling as the variable names.
      */
     /*!@{*/
-    std::vector<double>                          mpx;    ///< @brief @f$ p_{x,{\rm miss}} = -\sum p_{x} @f$
-    std::vector<double>                          mpy;    ///< @brief @f$ p_{y,{\rm miss}} = -\sum p_{x} @f$
-    std::vector<double>                          sumet;  ///< @brief @f$ \sigma E_{\rm T} = \sum p_{\rm T} @f$
+    std::vector<float>                           mpx;    ///< @brief @f$ p_{x,{\rm miss}} = -\sum p_{x} @f$
+    std::vector<float>                           mpy;    ///< @brief @f$ p_{y,{\rm miss}} = -\sum p_{x} @f$
+    std::vector<float>                           sumet;  ///< @brief @f$ \sigma E_{\rm T} = \sum p_{\rm T} @f$
     std::vector<std::string>                     name;   ///< @brief Name of MET contribution
     std::vector<MissingETBase::Types::bitmask_t> source; ///< @brief Source of MET contribution
     /*!@}*/
